Silent file descriptor checks
In order to avoid spam in journald, we just get the exit code and let
the checker output the error message.

Also, correct how we retrieve process in the healthcheck_port and _listen
functions.
"ss" doesn't allow to match some processes, like "neutron-l3-agent". We
therefore use the PID instead, provided by "pgrep".
The "-d" option of pgrep allow to prepare its output for the "grep -E",
preventing any need of a loop.

Change-Id: I1555a9b79c954e646fe9ae35272231c581cea03e
Closes-Bug: #1821782
Closes-Bug: #1821856