Stop dumb-init usage
The dumb-init is actually installed by fetching it from github and
putting it in the base image. We don't actually want that for our usage.
Additionally a new block that overrode the entrypoint in the mariadb
container which now causes the mysql container to fail if dumb-init is
not installed.

Change-Id: I2228a0da506daf207ea402d3e5a0f156b629862a
Closes-Bug: #1821970