Use ContainerImageRegistryCredentials for registry auth
This change passes the value of parameter
ContainerImageRegistryCredentials to the authenticate method in
container prepare. The value is a dict of dicts, so it is validated to
be in the expected format before being used. This allows prepare tasks
to pull or push from any registry which requires a username/password,
which is required by downstream.

Change-Id: I71b767d16a22c732cc18378cedf6c6599ccc5ae1
Closes-Bug: #1823579