Use a typemap file instead of symlinks for tags
To support manifest lists, the response for requesting a tagged
manifest will need to return different contents based on the
content-type requested. Representing a tag as a symlink to the
manifest directory is not flexible enough for this need.

This change switches over to a type-map[1] based implementation so
that future changes can add entries to different files based on the
content-type. Whenever any write operation is performed on an image,
any existing symlink based tags will be migrated to a typemap.

[1] https://httpd.apache.org/docs/2.4/mod/mod_negotiation.html#typemaps
Change-Id: I0e56e1d74ef8468f1b89da5329a8f1ced711c0f3