utils/config: generate param_config.json if param_config exists
In https://review.opendev.org/#/c/661093, we are preparing a new
interface "param_config" which exposes the parameters structure used by
the configuration tools.

To use this data, we want to write it on the host so it can be consummed
later by the configuration management tools (e.g. Hiera or Ansible
later?).
The file will be generated next to step_config.pp and will mostly be
useful when running Puppet on baremetal hosts for now:
A new hieradata file will be created in https://review.opendev.org/#/c/661093
called param_config.json, which will be at the top level hierarchy.
The "param_config" parameter will have a deep merge behavior so hiera
data files like extraconfig, etc, will be able to override this
configuration.

When Puppet is run _from_ the containers, the container-puppet.py will
read the same data directly from the puppet_config for each service (if
it does exist).

Change-Id: I8127203b82e503fb9d6013b47fdd3066aee1b474