Add the ability to compute osds number counting lvm devices
The previous version of this workbook only expected to find
the devices in a single list; with the new version of ceph-ansible
it's valid to pass the lvm_volumes, so we need to recompute the
total amount of osds.
This commit add a step that allows to get both devices and lvm_volumes
lists and calculate the len().

Closes-Bug: #1831458
Change-Id: Id5b0bb1874ea8ad327dd086cb1e48ff15d2974d7