Add ceph-ansible support to tripleo-ansible-inventory
The tripleo-ansible-inventory command produces an inventory
which works with TripleO but not with ceph-ansible. TripleO
ceph-ansible integration then creates its own inventory for
the ceph-ansible playbook run. This patch makes generating
a separate inventory for ceph-ansible no longer necessary.

TripleO services already map to the same inventory groups
used by ceph-ansible, e.g. ceph_clients:clients, ceph_mons:
mons, ceph_osds:osds, etc. This patch maps the names of the
TripleO inventory Ceph groups to the names of the inventory
groups used by ceph-ansible and creates a duplicate group
entry in the inventory, but with the names that ceph-ansible
expects.

Change-Id: Ie9880d7b6702b3c5f0e407e4ffe6f297912a1d10