tripleo-hieradata: introduce hieradata_per_host
Before this patch, the format of hieradata_content could only be formed an hash
where each key is a datafile and the value is the hieradata to write in this
datafile.

Example:

extraconfig:
  foo: bar
  foo2: bar2
other_hierafile:
  foo: bar3

With per-host hieradata, we are going to provide a map of hieradata
files per host.
hieradata_content can now to be formed an hash where each key is a host composed
of datafiles and the value for each datafile is the hieradata to write in it.

Example:

controller1:
  extraconfig:
    foo: bar
    foo2: bar2
  other_hierafile:
    foo: bar3
controller2:
  extraconfig:
    foo: bar2
    foo2: bar3
  other_hierafile:
    foo: bar

This is done via hieradata_per_host parameter, set to False by default
for backward compatibility.

Change-Id: I0556e788be1a898ab26934a660273bb6134f35e6