Increase timeout of temp swift URLs from 1 to 4 hours
It can take longer than 1 hour for a temporary swift
URL which stores the backup of the fetch directory to
be created vs used. Increase the timeout from 1 hour
to 4 hours so that the URL does not expire before it
is used.

Change-Id: Ib989bd673c694a9dc5af2a4a63ed84888f102a50
Fixes-Bug: #1832509
(cherry picked from commit 8f2e1064a2473a8218056ecfd932c714b11e0fe6)