Add template mode to tripleo-hieradata
Add a new tasks file (hieradata_vars.yaml) to the tripleo-hieradata role
that will create the hieradata files from jinja2 templates that are
included with the role.

The templates will be rendered based off of the values of
host_vars/group_vars/inventory_vars and variable files read out of Heat.

This functionality is a start to migrating the Hiera generation out of
Heat into pure Ansible.

Change-Id: I9ccccbfee44fc7c1375dec7357ccad6cba909c16