Add all_nodes template to tripleo-hieradata.
Add a template for the all_nodes hieradata to tripleo-hieradata.

The network hostnames are also added to the inventory as these vars are
needed by the template. The network IP's are already present in the
inventory. These variables will be used by various tasks that render
hiera data dynamically based on the Ansible inventory only.

Change-Id: I6e1e1c28dc09c9e04119db910068d62409a5afc8