Add tripleo-ansible directory namespace
The tripleo-ansible project will now install it's contents into a
namespaced directory. This was done to ensure we can smoothly
transition to using tripleo-ansible without having to worry about
conflicting files. To ensure our services are able to access this
namespace the configuration PATH has been updated to include the
new namespace.

Depends-On: Iedbf9d1034ee3396d6cb064c64169c21ebcf3e18
Change-Id: I128646937b02902f92c22259e97c9e6024769234
Signed-off-by: Kevin Carter <kecarter@redhat.com>