Handle registries with incorrect certs correctly
We actually end up allowing registries with incorrect certificates in
the upload code so they get treated the same as inscure registries.
This change updates the return code of the is_insecure_registry to match
how we handle these registries in the image uploader. Additionally, this
means that when the container image prepare code runs, we will correctly
include these registries that have ssl enabled but incorrect
certificates will correctly be included in the
DockerInsecureRegistryAddress setting.

Change-Id: I7fba881645ec2ea167c064be07ed6d4281b7ed3d
Closes-Bug: #1833751