Remove roles that have been imported into tripleo-ansible
These roles have all been imported into tripleo-ansible and can now be removed
from here. Several roles still remain in tripleo-common largely because they
have inflight changes which should be merged prior to doing the conversion.
A follow on PR will be made to clean-up the remaining roles when they're
ready.

The zuul job layout has been udpated to remove the "openstack-tox-molecule"
job. This was done because we no longer need it in this repo; all of
the roles have been moved to tripleo-ansible.

Change-Id: If7f9f2ce6f13d18fd864c475961275c611d84eb6
Signed-off-by: Kevin Carter <kecarter@redhat.com>