Add playbook to stop containers (stein only)
We need to stop several containers on the upgrades workflow.

This submission allows to reduce the duplicated code in THT.
It also handles the case in which we have podman as the CLI
but the containers running with docker.

The submission is proposed for tripleo-common only for
stable/stein, the master patch needs to be moved to the
tripleo-ansible repository as all the Ansible automation
is moved to that new repository.

Change-Id: I3c0580893aed5bda46d9104ffcee0f00cbabd9ee