Also write NetworkConfig to role dir
The NetworkConfig script is also written to the role directory within
the config-download directory, not just the individual server directory.
As we move towards making the NetworkConfig script generic for a role,
it can be consumed from the role dir by any node.

It doesn't matter at this point which node's NetworkConfig script we
write to the role dir, so it's just written whenever a node's script is
written. Once we've moved towards completely role generic Network
configuration, we can remove writing it to the individual server dir as
well.

Change-Id: I6bbca8eb256f387c4f74378baa5253890cfe1abf