wait_for_introspection_to_finish_error set status FAILED
Currently when wait_for_introspection_to_finish_error is reached,
the _introspect workflow will pass as a success even when it
actually failed. This means that the retry logic in the calling
workflow is never triggered.

wait_for_introspection_to_finish_error is hit in downstream CI
when ironic-introspector is under too much load to respond
to status poll requests.

This change also ensures callers to the
tripleo.baremetal.v1.introspect can override the default concurrency
so that this can be changed when required.

Change-Id: Ifd88ff9175bc6ca583e3826c59787680e25fbea3
Partial-Bug: #1836976