Fix swift-proxy healthcheck
With the current swift-proxy healthcheck the container status is
unhealty even the service is up.

The used ps/grep won't return the swift proxy processes as the
output gets truncated:

()[root@overcloud-controller-0 /]$ ps -e
    PID TTY          TIME CMD
      1 ?        00:00:00 dumb-init
      6 ?        00:02:52 swift-proxy-ser
     28 ?        00:00:25 swift-proxy-ser
   4753 pts/1    00:00:00 sh
   4797 pts/1    00:00:00 ps

As a result we end up in the else, checking the cache.

This moves the if to use pgrep instead.

Change-Id: I169b07553b171fcacb532bf58acc4200e270bfbb