Add reauthentication to sessions
This change allows a session to reauthenticate as needed. To accomplish this
a function object for authentication is set within the session which will
allow any method estabilishing a session to authenticate using the existing
credentials as needed. To ensure we're not breaking our workflow or
existing retry logic, conditions have been added to check for a 401 status
code. If a 401 is found, the session will now reauthenticate and raise an
HTTPError exception. This will allow our tenacity retry decorators to
provide for task retries, in the same way they have been, and to ensure
the way requests exceptions are raised are identical.

Change-Id: I3743b7212d5c5f0d4496ffb4134a1891587de01a
Signed-off-by: Kevin Carter <kecarter@redhat.com>