Add support for uploading local image
The current python image uploader only supports uploading an image from
a remote source to a repository. Since customers my have a custom image
locally and want to push it to an undercloud repository (non-docker
based), we need to expose a way to upload a local image to an undercloud
repository. This change adjusts the upload function to check if the
source image is local because it is prefixed with containers-storage:
rather than docker:// or no prefix.

Change-Id: Ie4f34f668bdb32395187662566d917e799249444
Related-Blueprint: podman-support