Fix dir when copying deployments.yaml
This patch corrects the backport that was done in commit
574ce9450f3148c9ab59d8f87a427555bd63396c.

Although the previous backport applied with no conflicts, it did not
actually work semanatically due to a refactoring from a different patch
(f8aa548ca692e330bdc47bf32b3f1c309e50d45c) that had not been backported
to queens.

The issue is now corrected in this patch by using the correct directory
(tmp_path) when copying deployments.yaml.

Change-Id: I423d48757569f6c8f07078d453486196d3174e7a