Honor trash_output when not using queue
Previously, trash_output was not honored if a queue was not being used
to post messages.

This patch changes the behavior so that trash_output will be honored
even if a queue is not being used, and all stdout/stderr will be
discarded.

Change-Id: I4fccfa0cb2a5382a52d63598f66dae446ff29c25
Closes-Bug: #1842102
(cherry picked from commit c7d44bc2ca67c40098982c88e9625aacd25b780d)