Use trash_output in create_admin_via_ssh workflow
When deploying a large amount of nodes, the create_admin_via_ssh
workflow could fail due to the large amount of ansible output generated.

This patch updates the tripleo.ansible-playbook action in the workflow
with trash_output:true so that the output is not saved in the mistral
DB.

There is a log file saved already in case the output is needed for debug
purposes.

Change-Id: I078b22fb0a0e7116f87419b444b8b4039db73ef8
Closes-Bug: #1842102