Introduce TripleoInventories object
This object can be constructed by passing it
a mapping of stack name to TripleoInventory
objects.

It has a merge method which merges all of the
TripleoInventory objects into one OrderedDict
which appends the stack name and an underscore
to each host group to avoid name collisions.

It has a write_static_inventory method which
writes the merged OrderedDict into an Ansible
inventory.

Change-Id: I522f1acbd39ad36ae1fec896188d3e20e0575566
Implements: blueprint split-control-plane-unified-inventory