healthcheck_port: run ss with both sudo & root as best effort
Privileged containers are running under the system pid namespace which
makes 'ss' output different from the container user used to run the
process.

e.g. nova-compute is run as nova on the overcloud, but the container is
privileged, so the previous patch with sudo didn't help to fix the
healthcheck. The 'ss' needs to be run as root, not as nova.

In this patch we run the ss twice, once as root, once with sudo, run
sort to make sure we get uniq output; then grep is as before.

Change-Id: Ia2897a6be3e000a9594103502b716431baa615b1
Co-Authored-By: Oliver Walsh <owalsh@redhat.com>
Related-Bug: #1843555