Ignore failure calling cell_v2_discover_hosts from deployment
This will fail when there is no nova database configured, as for a
nova-less undercloud. This call to cell_v2_discover_hosts was added to
make sure any recent auto-discovered hosts are mapped to a cell, so I
think ignoring all errors is acceptable in this case.

Related-Bug: #1742179
Change-Id: I77727bbee996d123aea442a76810dcee8e957a8f
Blueprint: nova-less-deploy