Warn when looking up grandparent resource name
When running config-download if the deployment name property was not set
on a deployment resource from a ResourceGroup, it will trigger the code
path to lookup the grandparent resource name.

This codepath can be very slow depending on the number of nodes in the
stack.

This patch adds a warning when this is happening so that it's clear that
the name property needs to be set on the deployment resource.

Change-Id: I2cf7acef0999e1c36d4baa8c793bdaa78bd441ad
(cherry picked from commit 8532a397def413872ead645c2350d8b34ef52bc4)
(cherry picked from commit af23433b2825982c575fe95330502c5abfee4913)