Write ansible log content to a different log file
This change moves the ansible log content to a different file so
we can have a more clear view of what is going on. Log data was
being ingested into python using the INFO handler but this was
resulting in non-sequential timestamps and multiple log formats
being used in the same file. To produce more easily interpreted
log files the log data is being separated. The python log file
will indicate when an ansible action is started and completed.
All ansible log data will be written to a single log file at,
the configured log handler file location, by default this is
`/var/log/tripleo-container-image-prepare-ansible.log`.

With these two log files we should be able to build a better
picture of everything going on within this tool-chain.

Change-Id: Ib208e53619901421d785ca0883f65954f02e59be
Signed-off-by: Kevin Carter <kecarter@redhat.com>