Improve ThreadPoolExecutor usage
This change addresses a few different issues with out ThreadPoolExecutor
usage.

Previously we were inefficiently looping on the job results which was a
blocking call and not asynchronous. This change switched the threadpool
usage to use as_completed so we'll handle jobs as they complete rather
than blocking until the the specific job we're looking at completes.

Additionally this switches to use a with statement for the executor so
the threads get cleaned up correctly when we're done with the block.

See example comments:
https://docs.python.org/3/library/concurrent.futures.html#threadpoolexecutor-example

The _inspect call was using a thread pool executor to run session calls
but was essentially a serial executing function so it has been updated
to remove the executor and just call the requests as needed.

Change-Id: Ia7abec997f4e503e1a2db82e05d4fe6e8696defc
Related-Bug: #1844446
(cherry picked from commit a1d89c7d6359a388f46dc3e71b16ebb11851f571)