Randomize the container list for uploads
When we work through the list of containers in an alphabetical fashion,
we end up duplicating much of the layer fetching because it can occur at
the same time. Things like cinder-api, cinder-backup, cinder-volume
share many of the same layers.  Since we don't ensure that we only do
a single fetching of a layer hash durring the multiprocessing, we end
up duplicating the fetches of layers. By randomizing the fetches, we
reduce the likelihood that we'll be fetching the same family of service
containers concurrently.

Change-Id: Ifbcd55de52c9e2283203b1c6e2adeb266d43eca6
Related-Bug: #1844446
(cherry picked from commit 3adfefa13aebd7bd9e48c952df9107d51a557011)