Prefix ansible output with timestamps
We output the ansible output in a single log line which can lead to what
appears to be out of sync timestamps.  We can split the output by line
and output that individually so the outter timestamps are in sync.

Change-Id: I02723e0900a7490cb980fe84fbeebe83137e8554
Closes-Bug: #1845930
(cherry picked from commit 6f92cb6685c75217606e74a3aaaf9702872513bf)