Raise exception when no IPs found
When generating the inventory, an exception is now raised if no IP
addresses are found for a given role from the specified network.
Previously, an unhandled IndexError was thrown with no indication as to
the actual problem.

This situtation can occur when using pre-provisioned nodes and a
malformed DeployedServerPortMap value such that no IP addresses are
assigned.

Change-Id: I2615bf4d5fcedbcf85d11e4aedbe22061348ded2