Handle empty cert related hostvar info
When certificates are neither generated nor provided (e.g. on update)
related variables would be empty or undefined and need to be handled
accordingly.

Note: this file has been relocated to triple-ansible in train+ so this
is a non-trivial backport of I361c90e438e79e180e2a745da82e619347fed13e.

Change-Id: I6afe5078bd285374b3a9311a066ba10883c2ff66
Closes-Bug: #1845366