Clean up instance defaults setting
This change fixes a problem where an instance entry was not being
populated with an image specified in the defaults dict. This change
sets up a baseline defaults dict before validating the roles, and
populates any instances with the defaults in-place.

This is a low-risk backport as the provision workflow is new and not
documented to be used in stable/train.

Change-Id: I1929e1b135ed539a6da09b1ec5d2b64b6281dc47
Blueprint: nova-less-deploy
Closes-Bug: 1851575
(cherry picked from commit 82f3a35fdce9a9ab9a63295afabd91fa7761f60a)