Re-raise same exception type in export_stream
tripleo-container-image-prepare can consume a lot of memory during
the export. Previously, when a MemoryError was raised by python, the
exception was hidden because there was no message in the exception and
the code re-raised an IOError.

This patch updates the code so that a MemoryError is specifically caught
if one was thrown, and an accurate message is logged.

Change-Id: I6de47600f430d3af5e3052ff456cd718a0df0d76
(cherry picked from commit 9d7a9b83ed8d20c8baa9a6b49812f4f30faf7782)