Revert "Set display_skipped_host to no"
display_skipped_hosts=no has the unfortunate side effect of not showing
the running task name until the task has returned. The reason is b/c the
task name won't be shown at all if all hosts are skipped, but that won't
be known until the task returns.

For long running tasks, this can be quite misleading as the running task
name is not seen until it returns. This makes debugging difficult, and users
don't know what task is actually running until it's completed.

There is an option to show the task name as soon as it starts executing,
but it was not added until ansible-2.9:
https://github.com/ansible/ansible/pull/53819

See also https://github.com/ansible/ansible/issues/51042

This reverts commit 403535c5722cda7cdc7573b009046fb0c2f86c7c.

Change-Id: Ia9e291d476d3143b194ce06002a0c1745cf3ef94