Remove checking for already running config_download_deploy
This patch removes the checking for already running instances of the
tripleo.deployment.v1.config_download_deploy workflow from the workflow
itself as it's been moved to tripleoclient.

It is better handled in tripleoclient where the workflow inputs can be
checked to see if any of the other running workflows are also for the
current stack. That functionality would have required a custom action to
do in the workflow itself.

Change-Id: I6195068a42bfc2469a0b8b006e339e3ca5056dff
Partial-Bug: #1852314
Depends-On: Ic8dbf28b5796ff998165b6b73b941f21c65f1dfa