Add RootStackName parameter on stack update
The task to add the RootStackName parameter to plan-environment.yaml in
the plan was only in the create_deployment_plan workflow. This patch
also adds it to the update_deployment_plan workflow so that the
parameter will be present in the environment for both stack create and
update.

Change-Id: I560eaff21c22d2ab70c657b35a3d43a76003e6ba
Closes-Bug: #1853362
(cherry picked from commit 07444eed0736ceab8af6ca7b37877223cfc4ade7)