Provide container name when raising exceptions
Currently exceptions thrown during image builds only output the
Future object as a string. This is not very helpful as it does
not give any information on what container build threw the
exception. This patch uses the mapping created between the
Futures and container names to provide more detailed information.

Change-Id: Idb3e2ce893f8e493144ab9101fec9a13b375cf33