Pass container_config to config actions
This patch updates the get_config and update_nodes workflows to pass in
the container_config name into the get_overcloud_config/download_config
actions.

When the container_config name is not passed, the default
overcloud-config is used. This means the correct config is not used if
the stack name is not overcloud as is often the case with multiple
overclouds or multi-stack.

Change-Id: Id6b4a5913141fc558c40f2cfd29f5e0741544ab4
Closes-Bug: #1854991
(cherry picked from commit 12ecf6c85d6db28794e5ce04c2c8a29d6c9fc53c)