config: refactor how per step tasks are generated
The per step tasks generation was introduced by:
I4d864f374d6f840585fafef2c7678e55c154898e

This patch is refactoring a little bit so we can easily re-use that
interface for the other tasks.
It introduces a new constant: PER_STEP_TASKS
It's a list of tasks that are "per step" ready.

Note about the 'else' in tripleo_common/utils/config.py :
Once all tasks are adapted in THT to run per step, we will be able to move this
condition to the upper level
We include it here to allow the CI to pass until THT changed is not merged.

Change-Id: Ie03084bb599b7b06aeeb321d2a7938a908487788