Change the python interpreter discovery mode.
Current default mode for the python interpreter discover inansible 2.8 is
auto_legacy. This patch changes the mode to auto, the biggest difference
respecting auto_legacy is 'If no entry is found, or the listed Python is
not present on the target host, searches a list of common Python interpreter
paths and uses the first one found' [0].
Currently, it has been observed some issues with the
discovered_python_interpreter fact not getting updated on specific scenarios
(for example, when upgrading the node from RHEL7 to RHEL8). This change is
expected to improve this situation.

[0] - https://docs.ansible.com/ansible/latest/reference_appendices/interpreter_discovery.html#
Related-Bug: #1856313
Change-Id: Iaef4839bb15ec398537b5c57a441c8e28a552bc0