workbooks/deployment: ignore uuid to hostname errors with splitstack & nova-less
The lookup that converts a Nova server UUID into a node hostname from
Ironic doesn't work if the baremetal is managed in split-stack &
nova-less.

For the time being we address these 2 use cases, let's just skip Ansible
scale down tasks for these deployments.

A proper lookup will be provided in the future and backported down to
stable/train.

Change-Id: I90f7f5a8966a6942e5dcfeb753996df420665be7
Closes-Bug: #1858459