image_uploader (attempt #3): fix images upload with no labels
If an image has no label, we set labels to {}; so to build tag_label in
that case we need to catch the TypeError exception or
tag_from_label.format(**labels) will raise, since labels would be NoneType.

We could have remove the default {} for Labels; but it's better to keep
it for further use in the image uploader; when the parameter is required
for certain methods.

Closes-Bug: #1857012
Change-Id: I35d73e7eca6f3cc208eda5d4c78a7bdd6cd7b810